Cells extracting strategy, and adding implementation of line grid cutter

+++ Web/RecognizerLib/src/main/java/pl/sudokusolver/recognizerlib/ocr/IRecognizer.java
@@ -3,5 +3,5 @@ package pl.sudokusolver.recognizerlib.ocr;
 import org.opencv.core.Mat;
 
 public interface IRecognizer {
-    short detect(Mat img);
+    int detect(Mat img);
 }
