+++ Web/RecognizerLib/src/main/java/pl/sudokusolver/recognizerlib/ocr/ml/KNN.java
@@ -14,11 +14,11 @@ public class KNN extends MLWrapper {
         sampleSize = data.getSize();
     }
 
-    public short detect(Mat img) {
+    public int detect(Mat img) {
         Mat wraped = applyFilter(img);
         Mat result = new Mat();
 
         knn.findNearest(ImageProcessing.procSimple(wraped, sampleSize), 3, result);
-        return (short) result.get(0,0)[0];
+        return (int) result.get(0,0)[0];
     }
 }
